Adding hover and approved states
This update includes new map functionality that allows users to see more clearly what zone polygon they're interacting with by showing a hover state when the mouse cursor enters a polygon. 

It also includes new styling that renders polygons with the property "Status" with the state "Approved" in a different color.